Updated read me Tracker section
I have updated the tracker section to identify where the trackers are to be used. For example in the www2.gov.bc.ca tracker there is the node id context that will not allow the tracker to work on other sites. This is also true for the search gov tracker. The engage tracker has the search tracking but it uses the search parameter that is only on the engage sites.